add: setup of scraper, model and mongodb
