+++ gsp/gsp/mongodb/setup.py
@@ -1,14 +1,12 @@
-import datetime
 import os
 from dotenv import load_dotenv
 from pymongo import MongoClient
 from pymongo.database import Database
 from pymongo.collection import Collection
 from lib.logger.setup import setup_logger
-from gsp.utils.model_utils import get_nth_previous_working_date
-from data import load_output
 
-logger = setup_logger()
+
+logger = setup_logger(__name__)
 
 
 def setup_mongodb() -> MongoClient:
@@ -87,36 +85,3 @@ def setup_collection(database: Database, collection_name: str) -> Collection:
         collection = database.get_collection(collection_name)
 
     return collection
-
-
-def run():
-    logger.info("Setting up MongoDB connection...")
-    client = setup_mongodb()
-
-    logger.info("Setting up MongoDB database...")
-    database = setup_database(client)
-
-    logger.info("Setting up MongoDB collections...")
-    generations_collection = setup_collection(database, "generations")
-    predictions_collection = setup_collection(database, "predictions")
-
-    logger.info("Loading dataframes...")
-    prediction_date = get_nth_previous_working_date(n=0, date=datetime.date.today())
-    generation_df = load_output(f"generation_{prediction_date.isoformat()}.csv")
-    prediction_df = load_output(f"prediction_{prediction_date.isoformat()}.csv")
-
-    # logger.info("Transforming dataframes into mongodb documents...")
-    # generations_doc = generation_df.to_dict(orient="records")
-    # predictions_doc = prediction_df.to_dict(orient="records")
-
-    # logger.info("Inserting documents into MongoDB collections...")
-    # generation should have a list of references to predictions
-    # generations_collection.insert_many(generations_doc)
-    # predictions_collection.insert_many(predictions_doc)
-
-    logger.info("Closing MongoDB connection...")
-    client.close()
-
-
-if __name__ == "__main__":
-    run()
