+++ gsp/pyproject.toml
@@ -29,7 +29,7 @@ pymongo = "^4.7.2"
 python-dotenv = "^1.0.1"
 
 [tool.poe.tasks]
-scrape = "python -m gsp.scraper.run"
+scrape = "python -m gsp.scraper.scraper"
 lint = "flake8 ."
 fmt = "black ."
 fmt-check = "black --check ."
